requirements.txt
pyyaml==3.13
python-dateutil==2.8.0
pytest==4.0.1
boto3==1.9.82
pandas==0.23.4
scikit-learn==0.20.3
numpy == 1.14.2
xgboost == 0.82
flask == 1.0.2
flask-sqlalchemy  == 2.3.2
SQLAlchemy==1.3.1